Remove async_io, simplify AsyncModbus<x>Client.
