--- doc/source/library/pymodbus.client.asynchronous.async_io.rst
@@ -1,8 +1,0 @@
-pymodbus\.client\.asynchronous\.async_io package
-================================================
-
-.. automodule:: pymodbus.client.asynchronous.async_io
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
