+++ doc/source/library/pymodbus.client.rst
@@ -1,7 +1,7 @@
 pymodbus\.client package
 ========================
 
-Pymodbus offers a :mod:`synchronous client <pymodbus.client.sync>`, and async clients based on :mod:`asyncio <pymodbus.client.asynchronous.async_io>` and :mod:`Twisted <pymodbus.client.asynchronous.Twisted>`.
+Pymodbus offers a :mod:`synchronous client <pymodbus.client.sync>`, and async clients based on :mod:`asyncio <pymodbus.client.asynchronous>`.
 
 Each client shares a :mod:`common client mixin <pymodbus.client.common>` which offers simple methods for reading and writing.
 
@@ -30,5 +30,3 @@ pymodbus\.client\.sync module
     :members:
     :undoc-members:
     :show-inheritance:
-
-
